social/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
python-dotenv==0.16.0
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==4.6.0
Pillow==2.2.1
psycopg2-binary==2.8.5
